Preserve 'use_cache' and 'validate' in context
Before this fix we were losing information in the context when
reindexing and the dictization_functions validate method was
causing the product theme tags to be moved from tags in the
dataset dict to 'product_themes'. This was causing problems
when Solr would complain about 'product_themes' not being
multivalue.

With this fix, product_themes are correctly stored as tags
in the dataset dict so they get converted to vocab_* fields
in Solr (which is mutivalue) and Solr doesn't complain